QA_System/requirements.txt
# 基础依赖
pyyaml>=6.0
numpy>=1.20.0
scikit-learn>=1.0.0
tqdm>=4.64.0

# 自然语言处理相关
nltk>=3.6.0
rank_bm25>=0.2.2

# Web UI相关
gradio>=3.33.0

# 评估相关
pandas>=1.3.0

# 其他工具
requests>=2.28.0

# DPR模型训练相关
torch>=2.0.0
transformers>=4.28.0
faiss-cpu>=1.7.0
tensorboard>=2.12.0  # 用于训练可视化
sentencepiece>=0.1.99  # 用于DPR分词 

# 其他依赖
langdetect>=1.0.9
sentence-transformers>=2.2.0
pytest>=7.0.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
langchain>=0.0.208
unstructured>=0.7.0

python main.py --config config.yaml --verbose